Tweak block hardness, change to return float seconds
Now that GH-106 block break animation is visible, greater block hardness
is possible since there is feedback the user is making progress mining.

+++ src/item.h
@@ -71,6 +71,6 @@ bool is_obstacle(int w);
 bool is_transparent(int w);
 bool is_destructable(int w);
 int is_illuminated(int w);
-int is_hardness(int w);
+float is_hardness(int w);
 
 #endif
